Move text editor to /texts/
